Url-Shortener | Container Components Separating components into container and presentational components. Renaming component.
