+++ app/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Grid } from 'react-bootstrap';
-import LinkInput from './Link-Input/LinkInput';
-import LinkList from './Link-List/LinkList';
+import LinkSubmit from './Link-Submit/LinkSubmitContainer';
+import LinkList from './Link-List/LinkListContainer';
 import Header from './Header/Header';
 
 const App = () => (
@@ -9,7 +9,7 @@ const App = () => (
 
     <Header />
 
-    <LinkInput />
+    <LinkSubmit />
 
     <LinkList />
