+++ app/components/Link-Submit/LinkSubmit.js
@@ -1,10 +1,8 @@
 import React, { Component } from 'react';
 import { Row, Col } from 'react-bootstrap';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { createShortenUrl } from '../../actions/linkSubmitActions';
 
-class LinkInput extends Component {
+class LinkSubmit extends Component {
   constructor(props) {
     super(props);
 
@@ -52,12 +50,12 @@ class LinkInput extends Component {
               value={url}
               onChange={this.onChangeUrl}
               name="url"
-            />
+              />
             <button
               className={`button button--primary form__button ${buttonClass}`}
               type="submit"
               disabled={!url}
-            >
+              >
               Shorten this link
             </button>
           </form>
@@ -66,17 +64,10 @@ class LinkInput extends Component {
     );
   }
 }
-const mapStateToProps = state => ({
-  isLoading: state.submitIsLoading,
-});
 
-const mapDispatchToProps = dispatch => ({
-  onCreateShortenUrl: url => dispatch(createShortenUrl(url)),
-});
-
-LinkInput.propTypes = {
+LinkSubmit.propTypes = {
   isLoading: PropTypes.bool.isRequired,
   onCreateShortenUrl: PropTypes.func.isRequired,
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(LinkInput);
+export default LinkSubmit;
